added the float cast again. maybe the compiler thought it was unnecessary, but actually without it, the private text was no longer centered around the player.

+++ src/games/stendhal/client/entity/Text.java
@@ -78,7 +78,7 @@ public class Text extends Entity {
 		} else {
 			this.tx = x
 					+ 0.7
-					- (textImage.getWidth() / GameScreen.SIZE_UNIT_PIXELS * 2.0f);
+					- (textImage.getWidth() / ((float) GameScreen.SIZE_UNIT_PIXELS * 2.0f));
 			this.ty = y + 1.5;
 		}
